move reactive types into own module
